microservices/deployment/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
grpcio==1.68.0
h11==0.14.0
idna==3.10
numpy==2.1.3
protobuf==5.28.3
pydantic==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
